Improvement escapes of vars

+++ app/models/UsersManager.php
@@ -5,6 +5,7 @@ namespace Laurent\App\Models;
 use Laurent\App\Models\Model;
 use Laurent\App\Models\Users;
 use Laurent\App\Service\Security;
+use Laurent\App\Service\Flash;
 
 
 class UsersManager extends Model
@@ -33,7 +34,7 @@ class UsersManager extends Model
 			isset($_POST['regist_firstname']) &&
 			isset($_POST['regist_email']) &&
 			isset($_POST['regist_username']) &&
-			isset($_POST['regist_password']) &&
+			isset($_POST['regist_password'])&&
 			isset($_POST['regist_password_confirm']))
 		{
 			$request = $this->_db->prepare('INSERT INTO users (lastname, firstname, email, username, password, rank, confirmation_token) VALUES(?, ?, ?, ?, ?, ?, ?)');
@@ -104,31 +105,10 @@ class UsersManager extends Model
 		return $user;
 	}
 
-    /**
-     * Méthode permettant d'enregistrer une news.
-     * @param Users $users
-     * @return void
-     * @see self::add()
-     * @see self::update()
-     */
-	public function save(Users $users)
-	{
-	  if ($users->isValable())
-	  {
-	    $users->isNew() ? $this->add($users) : $this->update($users);
-	  }
-	  else
-	  {
-	    throw new RuntimeException('La news doit être valide pour être enregistrée');
-	  }
-	}	
-
 	public function getUserInfos($id)
 	{
 		$request = $this->_db->prepare('SELECT * FROM users WHERE id = ?');
-			$request->execute(array($id));
-
-			// $request->setFetchMode(\PDO::FETCH_CLASS | \PDO::FETCH_PROPS_LATE, 'Users');
+		$request->execute(array($id));
 
 		$data = $request->fetch();
 		$userInfos = new Users($data);
@@ -234,8 +214,6 @@ class UsersManager extends Model
 		$request = $this->_db->prepare('SELECT * FROM users WHERE id = ? AND confirmation_token = ?');
 		$request->execute([$user_id, $token]);
 
-		// $request->setFetchMode(\PDO::FETCH_CLASS | \PDO::FETCH_PROPS_LATE, 'Users');
-
 		$data = $request->fetch();
 		$user = new Users($data);
 
@@ -247,7 +225,6 @@ class UsersManager extends Model
 		$user_id = htmlspecialchars($_GET['user_id']);
 
 		$request = $this->_db->prepare('UPDATE users SET confirmation_token = NULL, confirmedAt = NOW() WHERE id = ?')->execute([$user_id]);
-		// $request->closeCursor();
 	}
 
 	public function checkMailExistance()
@@ -266,25 +243,31 @@ class UsersManager extends Model
 	public function makeConnexionOfUser()
 	{
 		$user = $this->checkUser();
+		
 		if($user != NULL && !isset($_SESSION['auth']))
 		{
-			$this->_security = new Security;
-			//defuse the bruteforce checking
-			$this->deleteIp($this->_security->getIp());
-			// Initialisation des datas
+			if($user->confirmedAt() != NULL)
+			{
+				$this->_security = new Security;
+				//defuse the bruteforce checking
+				$this->deleteIp($this->_security->getIp());
+				// Initialisation des datas
 
-			$_SESSION['lastTime'] = time();
+				$_SESSION['lastTime'] = time();
 
-			$_SESSION['auth'] = 1;
-			$_SESSION['id'] = $user->id();
-			$_SESSION['lastname'] = $user->lastname();
-			$_SESSION['firstname'] = $user->firstname();
-			$_SESSION['email'] = $user->email();
-			$_SESSION['username'] = $user->username();
-			$_SESSION['password'] = $user->password();
-			$_SESSION['rank'] = $user->rank();
-
-			$_SESSION['tokenCsrf'] = md5(time()*rand(1,1000));	
+				$_SESSION['auth'] = 1;
+				$_SESSION['id'] = $user->id();
+				$_SESSION['lastname'] = $user->lastname();
+				$_SESSION['firstname'] = $user->firstname();
+				$_SESSION['email'] = $user->email();
+				$_SESSION['username'] = $user->username();
+				$_SESSION['password'] = $user->password();
+				$_SESSION['rank'] = $user->rank();
+				$_SESSION['tokenCsrf'] = md5(time()*rand(1,1000));
+			}
+			FLASH::setFlash('Validez votre adresse email avant votre première connexion', 'success');
+			header('Location: connexion');
+			exit();
 		}
 	}
 
@@ -296,7 +279,6 @@ class UsersManager extends Model
 	public function getUsersList()
 	{
 		$request = $this->_db->query('SELECT id, lastname, firstname, email, username, password, rank, DATE_FORMAT(confirmedAt, \'%d/%m/%Y à %Hh%i\') AS confirmedAt FROM users');
-		// $request->setFetchMode(\PDO::FETCH_CLASS | \PDO::FETCH_PROPS_LATE, 'Users');
 
 		$allUsers = [];
 
@@ -327,7 +309,17 @@ class UsersManager extends Model
 		$request->bindValue(':id', $users->id(), \PDO::PARAM_INT);
 		$request->bindValue(':rank', 2);
 
-		// var_dump($users->id()); die();
+
+		$request->execute();
+		$request->closeCursor();
+	}
+
+	public function downgradeUser(Users $users)
+	{
+		$request = $this->_db->prepare('UPDATE users SET rank = :rank WHERE id = :id');
+
+		$request->bindValue(':id', $users->id(), \PDO::PARAM_INT);
+		$request->bindValue(':rank', 1);
 
 		$request->execute();
 		$request->closeCursor();
